# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: AIPY API
  version: 0.1.0
  description: |
    MVP contract draft for the multi-tenant content workflow platform.

    WARNING: this document is hand-written and does NOT match the implementation.
    It still uses `/session`, `/content-runs` and `/decisions`, while the running
    app exposes `/me`, `/workflow-runs` and `/actions`. The authoritative route
    list is whatever `create_app().openapi()` returns; regenerate before relying
    on this file.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      security: []
      tags: [auth]
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
      responses:
        "200":
          description: Token pair.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "401":
          $ref: "#/components/responses/Problem"
  /auth/refresh:
    post:
      security: []
      tags: [auth]
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        "200":
          description: Rotated token pair.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Problem:
      description: Stable problem response.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    TokenPair:
      type: object
      required: [access_token, refresh_token, access_expires_at, refresh_expires_at]
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        access_expires_at: { type: string, format: date-time }
        refresh_expires_at: { type: string, format: date-time }
...
    Problem:
      type: object
      required: [type, title, status, code, request_id]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        detail: { type: [string, "null"] }
        request_id: { type: string }